Distiller/requirements.txt
click==6.7
Flask==0.12.2
Flask-SocketIO==2.9.2
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pip==9.0.1
python-engineio==1.7.0
python-socketio==1.8.0
RPi.GPIO==0.6.3
setuptools==28.8.0
six==1.10.0
SQLAlchemy==1.1.13
Werkzeug==0.12.2
